feat(storybook-pages): add Angular stories for Justifi web components (#172)
This commit adds a new file `angular.stories.mdx` to the `stencil-library/storybook-pages` directory. The file contains documentation and examples for using Justifi web components with Angular.

The documentation includes sections on usage, usage details, and styling. It provides instructions for installing the Angular version of the web components, importing them into a module, and using them as Angular components. It also explains how to handle props names, listen to events, and call methods in Angular. Additionally, it provides information on styling the components using CSS variables and styling exported parts.

This documentation aims to help users integrate Justifi's Angular web components and provides a starting point for using them in Angular projects.