refact: re-write PartsTable and PropsTable to not use React
Removed the React dependecy on the PartsTable and PropsTable helper functions.

These helper functions now return only HTML.

Updated the files that uses these helpers.

Escaped apostrophes on strings to avoid breaking the compiler on the public-docs site. Acorn library breaks when it tries to parse a string with unescaped apostrophe.